+++ util.hpp
@@ -1,5 +1,5 @@
 #include <string>
 #include <vector>
 
-std::vector<int> readIntegers(const std::string& fileName);
+std::vector<int64_t> readIntegers(const std::string& fileName);
 std::vector<std::string> readStrings(const std::string& fileName);
